SecretEntrance "constructor" and 'zeroes' and 'clicks' methods.
I like the idea (forced on me here) to have the data as what's
passed around, and a module specific to that data (format)
to manipulate it. For AdventOfCode exercises, it's a trivial
adjustment to make because we only work with a list of data (text)
and need only two module methods to (reduce) them to the answer.

Looks like the pattern hereout will be an 'Enum.map' for the initial
data ingest, and then an 'Enum.reduce' to whittle our way to truth.